+++ gdcef/src/utils.rs
@@ -105,3 +105,22 @@ pub fn get_subprocess_path() -> CefResult<PathBuf> {
         .canonicalize()
         .map_err(CefError::from)
 }
+
+/// Attempts to acquire a mutex lock, logging a warning on failure.
+macro_rules! try_lock {
+    ($mutex:expr, $context:literal) => {
+        match $mutex.lock() {
+            Ok(guard) => Some(guard),
+            Err(e) => {
+                godot::global::godot_warn!(
+                    "[CefTexture] Failed to acquire lock for {}: {}",
+                    $context,
+                    e
+                );
+                None
+            }
+        }
+    };
+}
+
+pub(crate) use try_lock;
